move front end source files to web_src and generate index.js minify version
